added deletion validation for expenses

+++ app/src/main/java/pl/edu/agh/iisg/to/to2project/app/expenses/accounts/controller/DeleteAccountPopupController.java
@@ -2,6 +2,7 @@ package pl.edu.agh.iisg.to.to2project.app.expenses.accounts.controller;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.text.Text;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Controller;
@@ -19,14 +20,21 @@ public class DeleteAccountPopupController extends PopupController {
 
     @Autowired
     private AccountService accountService;
-
     private Account account;
+
+    @FXML
+    public Text errorLabel;
 
     @FXML
     @Override
     protected void handleOKButtonClick(ActionEvent actionEvent) {
-        accountService.remove(account);
-        closeDialog();
+        if(accountService.canDelete(account)) {
+            accountService.remove(account);
+            closeDialog();
+        }
+        else{
+            errorLabel.setText("Sorry, you can't delete this account.");
+        }
     }
 
     public void deleteAccount(Account account){
